fix issues

+++ Core/include/Acts/Seeding/SeedFilterConfig.hpp
@@ -61,11 +61,11 @@ struct SeedFilterConfig {
   float minImpactSeedConf = 1. * Acts::UnitConstants::mm;
 
   // maximum number of lower quality seeds in seed confirmation
-  float maxSeedsPerSpMConf = std::numeric_limits<float>::infinity();
+  int maxSeedsPerSpMConf = std::numeric_limits<float>::infinity();
   // maximum number of quality seeds for each middle-bottom SP-duplet in seed
   // confirmation if the limit is reached we check if there is a lower quality
   // seed to be replaced
-  float maxQualitySeedsPerSpMConf = std::numeric_limits<float>::infinity();
+  int maxQualitySeedsPerSpMConf = std::numeric_limits<float>::infinity();
 
   // use deltaR between top and middle SP instead of top radius to search for
   // compatible SPs
